requirements.txt
fastapi[all]
sqlalchemy
alembic
python-jose[cryptography]
passlib
bcrypt==4.0.1
google-generativeai
pillow
python-dotenv
httpx
google-auth-oauthlib
authlib
starlette
python-multipart
itsdangerous
uvicorn
watchfiles
fastapi-mail